+++ input_methods.h
@@ -6,7 +6,7 @@
 #define INPUT_METHODS_H
 
 unsigned int validate_input(const char* formatSpecifier, void* value, char message[]);
-unsigned int get_length();
-double get_starting_number();
+unsigned long long get_length();
+long double get_starting_number();
 
 #endif //INPUT_METHODS_H
